Handle missing config keys during checkpoint search
When queried for checkpoints of a root graph, the `config` object the
checkpointer receives will not have a `checkpoint_id` or `checkpoint_ns`
key. Currently, we reinterpret their absence as presence but with either
an empty string or the string "None", neither of which will find
anything when queried. One because we aren't saving the string "None,"
and the other because this index doesn't support querying for empty
strings on these fields (would need the INDEXEMPTY option on fields,
which RedisVL doesn't currently support and only recent versions of
RediSearch support).

This commit doesn't entirely solve the problem, however. Now, we are
searching for any checkpoint for the thread, not checkpoints for the
thread for the root graph. We need to store a sentinel value like
"__empty__" and then query for it when the keys are not present in
config.