feat: Support Redis Cluster (#35)
Add support for Redis Cluster in the Store implementations:

Detect when the client is connected to a Redis Cluster or Enterprise node
If using Cluster, use single-key commands instead of multi-key commands, including transactions.
Allow users to override cluster detection. This is useful with something like the Redis Enterprise proxy, which can be used from a non-cluster-aware client.
This is "Part One" because our lower-level dependency, RedisVL, also needs to be updated to support cluster-aware clients. However, with the changes in this PR, users who use Redis clusters through a proxy (with a non-cluster-aware client, e.g., Redis, not RedisCluster) can set cluster_mode=True so we change our multi-key command behavior.